+++ R/call_slack_internals.R
@@ -9,13 +9,13 @@
 #' @keywords internal
 #' @noRd
 #' @references https://api.slack.com/methods/conversations.list
-list_channels <- function(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"), types = "public_channel", exclude_archived = TRUE, ...) {
+list_channels <- function(token = Sys.getenv("SLACK_TOKEN"), types = "public_channel", exclude_archived = TRUE, ...) {
   with_pagination(
     function(cursor) {
       call_slack_api(
         "/api/conversations.list",
         .method = GET,
-        bot_user_oauth_token = bot_user_oauth_token,
+        token = token,
         types = types,
         exclude_archived = exclude_archived,
         limit = 1000,
@@ -33,13 +33,13 @@ list_channels <- function(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUT
 #' @keywords internal
 #' @noRd
 #' @references https://api.slack.com/methods/users.list
-list_users <- function(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"), ...) {
+list_users <- function(token = Sys.getenv("SLACK_TOKEN"), ...) {
   with_pagination(
     function(cursor) {
       call_slack_api(
         "/api/users.list",
         .method = GET,
-        bot_user_oauth_token = bot_user_oauth_token,
+        token = token,
         ...,
         .next_cursor = cursor
       )
@@ -62,18 +62,17 @@ list_users <- function(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_T
 #' @param link_names Passed to `link_names` parameter of `chat.postMessage` API
 #'
 #' @references https://api.slack.com/methods/chat.postMessage
-post_message <- function(
-                         txt,
+post_message <- function(txt,
                          channel,
                          emoji = "",
                          username = Sys.getenv("SLACK_USERNAME"),
-                         bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
+                         token = Sys.getenv("SLACK_TOKEN"),
                          ...) {
   z <-
     call_slack_api(
       "/api/chat.postMessage",
       .method = POST,
-      bot_user_oauth_token = bot_user_oauth_token,
+      token = token,
       body = list(
         text       = txt,
         channel    = channel,
@@ -105,17 +104,16 @@ post_message <- function(
 #'
 #'
 #' @references https://api.slack.com/methods/files.upload
-files_upload <- function(
-                         file,
+files_upload <- function(file,
                          channel,
                          txt = "",
                          username = Sys.getenv("SLACK_USERNAME"),
-                         bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
+                         token = Sys.getenv("SLACK_TOKEN"),
                          ...) {
   z <- call_slack_api(
     "/api/files.upload",
     .method = POST,
-    bot_user_oauth_token = bot_user_oauth_token,
+    token = token,
     body = list(
       file            = upload_file(file),
       initial_comment = txt,
@@ -128,11 +126,11 @@ files_upload <- function(
 }
 
 
-list_scopes <- function(bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")) {
+list_scopes <- function(token = Sys.getenv("SLACK_TOKEN")) {
   z <- call_slack_api(
     "/api/apps.permissions.scopes.list",
     .method = GET,
-    bot_user_oauth_token = bot_user_oauth_token
+    token = token
   )
   invisible(content(z))
 }