+++ R/slackr_csv.R
@@ -9,7 +9,8 @@
 #' @param channels Slack channels to save to (optional)
 #' @param title title on Slack (optional - defaults to filename)
 #' @param initial_comment comment for file on slack (optional - defaults to filename)
-#' @param bot_user_oauth_token Slack bot user OAuth token
+#' @param token A Slack token (either a user token or a bot user token)
+#' @param bot_user_oauth_token Deprecated. A Slack bot user OAuth token
 #' @param ... additional arguments to be passed to `write.csv()`
 #' @return `httr` response object from `POST` call (invisibly)
 #' @author Matt Kaye (aut)
@@ -21,8 +22,11 @@ slackr_csv <- function(data,
                        title = basename(filename),
                        initial_comment = basename(filename),
                        channels = Sys.getenv("SLACK_CHANNEL"),
+                       token = Sys.getenv("SLACK_TOKEN"),
                        bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
                        ...) {
+  token <- check_tokens(token, bot_user_oauth_token)
+
   write.csv(data, filename, ...)
 
   res <- slackr_upload(
@@ -30,7 +34,7 @@ slackr_csv <- function(data,
     title = title,
     initial_comment = initial_comment,
     channels = channels,
-    bot_user_oauth_token = bot_user_oauth_token
+    token = token
   )
 
   return(invisible(res))