+++ R/slackr_dev.R
@@ -4,7 +4,8 @@
 #' specified Slack channel.
 #'
 #' @param channels list of channels to post image to
-#' @param bot_user_oauth_token the Slack full bot user OAuth token (chr)
+#' @param token A Slack token (either a user token or a bot user token)
+#' @param bot_user_oauth_token Deprecated. A Slack bot user OAuth token
 #' @param file prefix for filenames (defaults to `plot`)
 #' @param plot_text the plot text to send with the plot (defaults to "")
 #' @importFrom grDevices dev.copy dev.off png
@@ -27,12 +28,11 @@
 #' }
 #' @export
 slackr_dev <- function(channels = Sys.getenv("SLACK_CHANNEL"),
-                       bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
+                       token = Sys.getenv("SLACK_TOKEN"),
                        plot_text = "",
-                       file = "plot") {
-  loc <- Sys.getlocale("LC_CTYPE")
-  Sys.setlocale("LC_CTYPE", "C")
-  on.exit(Sys.setlocale("LC_CTYPE", loc))
+                       file = "plot",
+                       bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")) {
+  token <- check_tokens(token, bot_user_oauth_token)
 
   ftmp <- tempfile(file, fileext = ".png")
   dev.copy(png, file = ftmp)
@@ -42,7 +42,7 @@ slackr_dev <- function(channels = Sys.getenv("SLACK_CHANNEL"),
     file = ftmp,
     channel = channels,
     txt = plot_text,
-    bot_user_oauth_token = bot_user_oauth_token
+    token = token
   )
 
   return(invisible(res))