+++ R/slackr_history.R
@@ -7,7 +7,8 @@
 #'   You need one or more of these scopes enabled in your slack app: *
 #'   channels:history * groups:history * im:history * mpim:history
 #'
-#' @param bot_user_oauth_token the Slack bot user OAuth token
+#' @param token A Slack token (either a user token or a bot user token)
+#' @param bot_user_oauth_token Deprecated. A Slack bot user OAuth token
 #' @param channel The channel to get history from
 #' @param posted_from_time Timestamp of the first post time to consider
 #' @param duration Number of hours of history to retrieve.  By default retrieves
@@ -21,14 +22,18 @@
 #' @return A `tibble` with message metadata
 #' @references <https://api.slack.com/methods/conversations.history>
 #'
-slackr_history <- function(
-                           channel = Sys.getenv("SLACK_CHANNEL"),
-                           bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
+slackr_history <- function(channel = Sys.getenv("SLACK_CHANNEL"),
+                           token = Sys.getenv("SLACK_TOKEN"),
                            posted_to_time = as.numeric(Sys.time()),
                            message_count,
                            duration,
                            posted_from_time,
-                           paginate = FALSE) {
+                           paginate = FALSE,
+                           bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")) {
+  token <- check_tokens(token, bot_user_oauth_token)
+
+  channel <- slackr_chtrans(channel, token)
+
   if (!missing(duration) && !is.null(duration) && !missing(posted_from_time) && !is.null(posted_from_time)) {
     posted_from_time <- posted_to_time - duration * 3600
   } else {
@@ -40,7 +45,7 @@ slackr_history <- function(
       resp <- call_slack_api(
         "/api/conversations.history",
         .method   = GET,
-        bot_user_oauth_token = bot_user_oauth_token,
+        token = token,
         channel   = channel,
         latest    = posted_to_time,
         oldest    = posted_from_time,
@@ -54,7 +59,7 @@ slackr_history <- function(
           call_slack_api(
             "/api/conversations.history",
             .method   = GET,
-            bot_user_oauth_token = bot_user_oauth_token,
+            token = token,
             channel   = channel,
             latest    = posted_to_time,
             oldest    = posted_from_time,