+++ R/slackr_save.R
@@ -7,7 +7,8 @@
 #' @param ... objects to store in the R data file
 #' @param channels Slack channels to save to (optional)
 #' @param file filename (without extension) to use
-#' @param bot_user_oauth_token Slack bot user OAuth token
+#' @param token A Slack token (either a user token or a bot user token)
+#' @param bot_user_oauth_token Deprecated. A Slack bot user OAuth token
 #' @param plot_text the plot text to send with the plot (defaults to "")
 #' @return `httr` response object from `POST` call
 #' @seealso [slackr_setup()], [slackr_dev()], [slackr_upload()]
@@ -21,13 +22,12 @@
 slackr_save <- function(...,
                         channels = Sys.getenv("SLACK_CHANNEL"),
                         file = "slackr",
-                        bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
-                        plot_text = "") {
-  if (channels == "") stop("No channels specified. Did you forget select which channels to post to with the 'channels' argument?")
+                        token = Sys.getenv("SLACK_TOKEN"),
+                        plot_text = "",
+                        bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")) {
+  token <- check_tokens(token, bot_user_oauth_token)
 
-  loc <- Sys.getlocale("LC_CTYPE")
-  Sys.setlocale("LC_CTYPE", "C")
-  on.exit(Sys.setlocale("LC_CTYPE", loc))
+  if (channels == "") abort("No channels specified. Did you forget select which channels to post to with the 'channels' argument?")
 
   ftmp <- tempfile(file, fileext = ".Rdata")
   save(..., file = ftmp, envir = parent.frame())
@@ -38,7 +38,7 @@ slackr_save <- function(...,
     file = ftmp,
     channel = channels,
     txt = plot_text,
-    bot_user_oauth_token = bot_user_oauth_token,
+    token = token,
     filename = sprintf("%s.Rdata", file)
   )