+++ R/slackr_tex.R
@@ -7,7 +7,8 @@
 #' @param channels list of channels to post image to
 #' @param ext character, type of format to return, can be tex, pdf, or any image device, Default: 'png'
 #' @param path character, path to save tex_preview outputs, if NULL then tempdir is used, Default: NULL
-#' @param bot_user_oauth_token the Slack full bot user OAuth token (chr)
+#' @param token A Slack token (either a user token or a bot user token)
+#' @param bot_user_oauth_token Deprecated. A Slack bot user OAuth token
 #' @param ... other arguments passed to [texPreview::tex_preview()], see Details
 #' @note You need to setup a full API token (i.e. not a webhook & not OAuth) for this to work
 #'       Also, you can pass in `add_user=TRUE` as part of the `...`
@@ -23,17 +24,15 @@
 #' @export
 slackr_tex <- function(obj,
                        channels = Sys.getenv("SLACK_CHANNEL"),
-                       bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
+                       token = Sys.getenv("SLACK_TOKEN"),
                        ext = "png",
                        path = NULL,
+                       bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
                        ...) {
+  token <- check_tokens(token, bot_user_oauth_token)
 
   # check if texPreview is installed, if not provide feedback
   check_tex_pkg()
-
-  loc <- Sys.getlocale("LC_CTYPE")
-  Sys.setlocale("LC_CTYPE", "C")
-  on.exit(Sys.setlocale("LC_CTYPE", loc))
 
   if (!is.null(path)) {
     td <- path
@@ -56,7 +55,7 @@ slackr_tex <- function(obj,
     add_headers(`Content-Type` = "multipart/form-data"),
     body = list(
       file = upload_file(file.path(td, paste0("slack.", ext))),
-      token = bot_user_oauth_token,
+      token = token,
       channels = channels
     )
   )
@@ -87,6 +86,6 @@ check_tex_pkg <- function() {
   )
 
   if (!is_installed) {
-    stop("texPreview package is not installed, run ?slackr_tex and see Details.", call. = FALSE)
+    abort("texPreview package is not installed, run ?slackr_tex and see Details.")
   }
 }