+++ R/slackr_upload.R
@@ -7,7 +7,8 @@
 #' @param title title on Slack (optional - defaults to filename)
 #' @param initial_comment comment for file on slack (optional - defaults to filename)
 #' @param channels Slack channels to save to (optional)
-#' @param bot_user_oauth_token Slack bot user OAuth token
+#' @param token A Slack token (either a user token or a bot user token)
+#' @param bot_user_oauth_token Deprecated. A Slack bot user OAuth token
 #' @return `httr` response object from `POST` call (invisibly)
 #' @author Quinn Weber (ctb), Bob Rudis (aut)
 #' @references <https://github.com/mrkaye97/slackr/pull/15/files>
@@ -18,16 +19,15 @@
 slackr_upload <- function(filename, title = basename(filename),
                           initial_comment = basename(filename),
                           channels = Sys.getenv("SLACK_CHANNEL"),
+                          token = Sys.getenv("SLACK_TOKEN"),
                           bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")) {
-  if (channels == "") stop("No channels specified. Did you forget select which channels to post to with the 'channels' argument?")
+  token <- check_tokens(token, bot_user_oauth_token)
+
+  if (channels == "") abort("No channels specified. Did you forget select which channels to post to with the 'channels' argument?")
   f_path <- path.expand(filename)
 
   if (file.exists(f_path)) {
     f_name <- basename(f_path)
-
-    loc <- Sys.getlocale("LC_CTYPE")
-    Sys.setlocale("LC_CTYPE", "C")
-    on.exit(Sys.setlocale("LC_CTYPE", loc))
 
     res <- POST(
       url = "https://slack.com/api/files.upload",
@@ -35,14 +35,14 @@ slackr_upload <- function(filename, title = basename(filename),
       body = list(
         file = upload_file(f_path), filename = f_name,
         title = title, initial_comment = initial_comment,
-        token = bot_user_oauth_token, channels = paste(channels, collapse = ",")
+        token = token, channels = paste(channels, collapse = ",")
       )
     )
 
-    if (!content(res)$ok) stop(content(res)$error, " -- Are you sure you used the right token and channel name?")
+    if (!content(res)$ok) abort(content(res)$error, " -- Are you sure you used the right token and channel name?")
 
     return(invisible(res))
   } else {
-    stop(sprintf("File [%s] not found", f_path), call. = FALSE)
+    abort(sprintf("File [%s] not found", f_path))
   }
 }