+++ tests/testthat/test-connection.R
@@ -7,15 +7,15 @@ test_that("Initial setup completes", {
         channel = Sys.getenv("SLACK_CHANNEL"),
         username = Sys.getenv("SLACK_USERNAME"),
         icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
-        incoming_webhook_url = Sys.getenv("SLACK_INCOMING_URL_PREFIX"),
-        bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN")
+        incoming_webhook_url = Sys.getenv("SLACK_INCOMING_WEBHOOK_URL"),
+        token = Sys.getenv("SLACK_TOKEN")
       )
     } else {
       slackr_setup(
         config_file = "~/.slackr_config"
       )
     },
-    'Successfully connected to Slack'
+    "Successfully connected to Slack"
   )
 })
 
@@ -25,8 +25,8 @@ test_that("slackr_setup() connects", {
   expect_equal(
     slackr_setup(
       channel = "#test",
-      bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
-      incoming_webhook_url = Sys.getenv("SLACK_INCOMING_URL_PREFIX"),
+      token = Sys.getenv("SLACK_TOKEN"),
+      incoming_webhook_url = Sys.getenv("SLACK_INCOMING_WEBHOOK_URL"),
       username = "slackr",
       icon_emoji = "robot_face"
     ),
@@ -34,11 +34,11 @@ test_that("slackr_setup() connects", {
   )
 
   expect_true(
-    grepl("^xox.-", Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"))
+    grepl("^xox.-", Sys.getenv("SLACK_TOKEN"))
   )
 
   expect_true(
-    grepl("^https://hooks", Sys.getenv("SLACK_INCOMING_URL_PREFIX"))
+    grepl("^https://hooks", Sys.getenv("SLACK_INCOMING_WEBHOOK_URL"))
   )
 
   expect_equal(
@@ -53,10 +53,10 @@ test_that("config file setup works", {
   tmp <- tempfile()
   write.dcf(
     list(
-      bot_user_oauth_token = Sys.getenv("SLACK_BOT_USER_OAUTH_TOKEN"),
+      token = Sys.getenv("SLACK_TOKEN"),
       channel = Sys.getenv("SLACK_CHANNEL"),
       username = Sys.getenv("SLACK_USERNAME"),
-      incoming_webhook_url = Sys.getenv("SLACK_INCOMING_URL_PREFIX"),
+      incoming_webhook_url = Sys.getenv("SLACK_INCOMING_WEBHOOK_URL"),
       icon_emoji = Sys.getenv("SLACK_ICON_EMOJI")
     ),
     file = tmp